# swagger.yaml
openapi: 3.0.0
info:
  title: API
  version: 1.0.0
servers:
  - url: http://localhost:3005
paths:
  /api/csv:
    post:
      tags:
        - default
      summary: csv
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /api/image:
    post:
      tags:
        - default
      summary: image
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /api/assessment:
    get:
      tags:
        - default
      summary: grades
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    post:
      tags:
        - default
      summary: insertAssessment
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                name: Rafał
                assessment: '4.0'
                date: '2024-04-22'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /api/assessment/{id}:
    put:
      tags:
        - default
      summary: putAssessment
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                name: Rafał
                assessment: '4.0'
                date: '2024-04-22'
      parameters:
        - name: id
          in: path
          schema:
            type: integer
          required: true
          example: '2'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}